requirements.txt
asgiref==3.5.2
autopep8==2.0.0
dj-rest-auth==2.2.5
Django==4.1.3
djangorestframework==3.14.0
pycodestyle==2.9.1
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
